+++ Editor/BuildSettings.cs
@@ -6,7 +6,7 @@ namespace HovelHouse.CloudKit
     {
         public bool EnablePostProcessBuild = true;
         public bool EnableCloudKit = true;
-        public bool EnableKeyVaueStorage;
+        public bool EnableKeyVaueStorage = true;
         public bool EnableDocumentStorage;
         public bool AddDefaultContainers = true;
         public string[] CustomContainers;