+++ Runtime/Plugin/CKAcceptSharesOperation.cs
@@ -1,7 +1,7 @@
 //
 //  CKAcceptSharesOperation.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -24,7 +24,6 @@ namespace HovelHouse.CloudKit
         // Class Methods
         
 
-        // Constructors
         
         #if UNITY_IPHONE || UNITY_TVOS
         [DllImport("__Internal")]
@@ -48,7 +47,6 @@ namespace HovelHouse.CloudKit
             );
         
 
-        // Instance Methods
         
 
         
@@ -82,17 +80,15 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern void CKAcceptSharesOperation_SetPropPerShareCompletionHandler(HandleRef ptr, PerShareCompletionDelegate perShareCompletionHandler, out IntPtr exceptionPtr);
         
+
         #endregion
 
         internal CKAcceptSharesOperation(IntPtr ptr) : base(ptr) {}
         
-        #region Class Methods
         
-        #endregion
-
-        #region Constructors
         
-        public static CKAcceptSharesOperation init(
+        
+        public CKAcceptSharesOperation(
             )
         {
             
@@ -105,11 +101,11 @@ namespace HovelHouse.CloudKit
                 throw new CloudKitException(nativeException, nativeException.Reason);
             }
 
-            return new CKAcceptSharesOperation(ptr);
+            Handle = new HandleRef(this,ptr);
         }
         
         
-        public static CKAcceptSharesOperation initWithShareMetadatas(
+        public CKAcceptSharesOperation(
             CKShareMetadata[] shareMetadatas
             )
         {
@@ -127,19 +123,14 @@ namespace HovelHouse.CloudKit
                 throw new CloudKitException(nativeException, nativeException.Reason);
             }
 
-            return new CKAcceptSharesOperation(ptr);
+            Handle = new HandleRef(this,ptr);
         }
         
         
-        #endregion
 
 
-        #region Methods
         
         
-        #endregion
-
-        #region Properties
         
         public CKShareMetadata[] ShareMetadatas 
         {
@@ -154,7 +145,7 @@ namespace HovelHouse.CloudKit
 
                 for (int i = 0; i < bufferLen; i++)
                 {
-                    IntPtr ptr2 = Marshal.ReadIntPtr(bufferPtr + (i * 8));
+                    IntPtr ptr2 = Marshal.ReadIntPtr(bufferPtr + (i * IntPtr.Size));
                     shareMetadatas[i] = ptr2 == IntPtr.Zero ? null : new CKShareMetadata(ptr2);
                 }
 
@@ -264,8 +255,9 @@ namespace HovelHouse.CloudKit
         }
 
         
-        #endregion
+
         
+
         
         #region IDisposable Support
         #if UNITY_IPHONE || UNITY_TVOS
@@ -277,10 +269,7 @@ namespace HovelHouse.CloudKit
             
         private bool disposedValue = false; // To detect redundant calls
         
-        // No base.Dispose() needed
-        // All we ever do is decrement the reference count in managed code
-        
-        private void Dispose(bool disposing)
+        protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
             {