+++ Runtime/Plugin/CKDatabaseOperation.cs
@@ -1,7 +1,7 @@
 //
 //  CKDatabaseOperation.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -17,17 +17,15 @@ using UnityEngine;
 
 namespace HovelHouse.CloudKit
 {
-    public class CKDatabaseOperation : CKOperation
+    public class CKDatabaseOperation : CKOperation, IDisposable
     {
         #region dll
 
         // Class Methods
         
 
-        // Constructors
         
 
-        // Instance Methods
         
 
         
@@ -48,25 +46,19 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern void CKDatabaseOperation_SetPropDatabase(HandleRef ptr, IntPtr database, out IntPtr exceptionPtr);
         
+
         #endregion
 
         internal CKDatabaseOperation(IntPtr ptr) : base(ptr) {}
+        internal CKDatabaseOperation(){}
         
-        #region Class Methods
         
-        #endregion
-
-        #region Constructors
         
-        #endregion
+        
 
 
-        #region Methods
         
         
-        #endregion
-
-        #region Properties
         
         public CKDatabase Database 
         {
@@ -81,8 +73,49 @@ namespace HovelHouse.CloudKit
             }
         }
         
-        #endregion
+
         
+
+        
+        #region IDisposable Support
+        #if UNITY_IPHONE || UNITY_TVOS
+        [DllImport("__Internal")]
+        #else
+        [DllImport("HHCloudKit")]
+        #endif
+        private static extern void CKDatabaseOperation_Dispose(HandleRef handle);
+            
+        private bool disposedValue = false; // To detect redundant calls
+        
+        protected override void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    // TODO: dispose managed state (managed objects).
+                }
+                
+                //Debug.Log("CKDatabaseOperation Dispose");
+                CKDatabaseOperation_Dispose(Handle);
+                disposedValue = true;
+            }
+        }
+
+        ~CKDatabaseOperation()
+        {
+            // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+            Dispose(false);
+        }
+
+        // This code added to correctly implement the disposable pattern.
+        public new void Dispose()
+        {
+            // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+        #endregion
         
     }
 }