+++ Runtime/Plugin/CKQuery.cs
@@ -1,7 +1,7 @@
 //
 //  CKQuery.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -24,7 +24,6 @@ namespace HovelHouse.CloudKit
         // Class Methods
         
 
-        // Constructors
         
         #if UNITY_IPHONE || UNITY_TVOS
         [DllImport("__Internal")]
@@ -48,7 +47,6 @@ namespace HovelHouse.CloudKit
             );
         
 
-        // Instance Methods
         
 
         
@@ -84,17 +82,15 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern IntPtr CKQuery_GetPropPredicate(HandleRef ptr);
         
+
         #endregion
 
         internal CKQuery(IntPtr ptr) : base(ptr) {}
         
-        #region Class Methods
         
-        #endregion
-
-        #region Constructors
         
-        public static CKQuery initWithRecordType(
+        
+        public CKQuery(
             string recordType, 
             NSPredicate predicate
             )
@@ -115,11 +111,11 @@ namespace HovelHouse.CloudKit
                 throw new CloudKitException(nativeException, nativeException.Reason);
             }
 
-            return new CKQuery(ptr);
+            Handle = new HandleRef(this,ptr);
         }
         
         
-        public static CKQuery initWithCoder(
+        public CKQuery(
             NSCoder aDecoder
             )
         {
@@ -136,19 +132,14 @@ namespace HovelHouse.CloudKit
                 throw new CloudKitException(nativeException, nativeException.Reason);
             }
 
-            return new CKQuery(ptr);
+            Handle = new HandleRef(this,ptr);
         }
         
         
-        #endregion
 
 
-        #region Methods
         
         
-        #endregion
-
-        #region Properties
         
         public NSSortDescriptor[] SortDescriptors 
         {
@@ -163,7 +154,7 @@ namespace HovelHouse.CloudKit
 
                 for (int i = 0; i < bufferLen; i++)
                 {
-                    IntPtr ptr2 = Marshal.ReadIntPtr(bufferPtr + (i * 8));
+                    IntPtr ptr2 = Marshal.ReadIntPtr(bufferPtr + (i * IntPtr.Size));
                     sortDescriptors[i] = ptr2 == IntPtr.Zero ? null : new NSSortDescriptor(ptr2);
                 }
 
@@ -203,8 +194,9 @@ namespace HovelHouse.CloudKit
             }
         }
         
-        #endregion
+
         
+
         
         #region IDisposable Support
         #if UNITY_IPHONE || UNITY_TVOS
@@ -216,10 +208,7 @@ namespace HovelHouse.CloudKit
             
         private bool disposedValue = false; // To detect redundant calls
         
-        // No base.Dispose() needed
-        // All we ever do is decrement the reference count in managed code
-        
-        private void Dispose(bool disposing)
+        protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
             {