+++ Runtime/Plugin/CKQueryNotification.cs
@@ -1,7 +1,7 @@
 //
 //  CKQueryNotification.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -24,10 +24,8 @@ namespace HovelHouse.CloudKit
         // Class Methods
         
 
-        // Constructors
         
 
-        // Instance Methods
         
 
         
@@ -55,25 +53,18 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern IntPtr CKQueryNotification_GetPropRecordID(HandleRef ptr);
         
+
         #endregion
 
         internal CKQueryNotification(IntPtr ptr) : base(ptr) {}
         
-        #region Class Methods
-        
-        #endregion
-
-        #region Constructors
-        
-        #endregion
-
-
-        #region Methods
         
         
-        #endregion
+        
 
-        #region Properties
+
+        
+        
         
         public CKDatabaseScope DatabaseScope 
         {
@@ -102,8 +93,9 @@ namespace HovelHouse.CloudKit
             }
         }
         
-        #endregion
+
         
+
         
         #region IDisposable Support
         #if UNITY_IPHONE || UNITY_TVOS
@@ -115,10 +107,7 @@ namespace HovelHouse.CloudKit
             
         private bool disposedValue = false; // To detect redundant calls
         
-        // No base.Dispose() needed
-        // All we ever do is decrement the reference count in managed code
-        
-        private void Dispose(bool disposing)
+        protected override void Dispose(bool disposing)
         {
             if (!disposedValue)
             {
@@ -140,7 +129,7 @@ namespace HovelHouse.CloudKit
         }
 
         // This code added to correctly implement the disposable pattern.
-        public void Dispose()
+        public new void Dispose()
         {
             // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
             Dispose(true);