+++ Runtime/Plugin/CKReference.cs
@@ -1,7 +1,7 @@
 //
 //  CKReference.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -24,7 +24,6 @@ namespace HovelHouse.CloudKit
         // Class Methods
         
 
-        // Constructors
         
         #if UNITY_IPHONE || UNITY_TVOS
         [DllImport("__Internal")]
@@ -49,7 +48,6 @@ namespace HovelHouse.CloudKit
             );
         
 
-        // Instance Methods
         
 
         
@@ -70,17 +68,15 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern IntPtr CKReference_GetPropRecordID(HandleRef ptr);
         
+
         #endregion
 
         internal CKReference(IntPtr ptr) : base(ptr) {}
         
-        #region Class Methods
         
-        #endregion
-
-        #region Constructors
         
-        public static CKReference initWithRecordID(
+        
+        public CKReference(
             CKRecordID recordID, 
             CKReferenceAction action
             )
@@ -99,11 +95,11 @@ namespace HovelHouse.CloudKit
                 throw new CloudKitException(nativeException, nativeException.Reason);
             }
 
-            return new CKReference(ptr);
+            Handle = new HandleRef(this,ptr);
         }
         
         
-        public static CKReference initWithRecord(
+        public CKReference(
             CKRecord record, 
             CKReferenceAction action
             )
@@ -122,19 +118,14 @@ namespace HovelHouse.CloudKit
                 throw new CloudKitException(nativeException, nativeException.Reason);
             }
 
-            return new CKReference(ptr);
+            Handle = new HandleRef(this,ptr);
         }
         
         
-        #endregion
 
 
-        #region Methods
         
         
-        #endregion
-
-        #region Properties
         
         public CKReferenceAction ReferenceAction 
         {
@@ -154,8 +145,9 @@ namespace HovelHouse.CloudKit
             }
         }
         
-        #endregion
+
         
+
         
         #region IDisposable Support
         #if UNITY_IPHONE || UNITY_TVOS
@@ -167,10 +159,7 @@ namespace HovelHouse.CloudKit
             
         private bool disposedValue = false; // To detect redundant calls
         
-        // No base.Dispose() needed
-        // All we ever do is decrement the reference count in managed code
-        
-        private void Dispose(bool disposing)
+        protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
             {