+++ Runtime/Plugin/CKShareMetadata.cs
@@ -1,7 +1,7 @@
 //
 //  CKShareMetadata.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -24,10 +24,8 @@ namespace HovelHouse.CloudKit
         // Class Methods
         
 
-        // Constructors
         
 
-        // Instance Methods
         
 
         
@@ -90,25 +88,18 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern CKShareParticipantRole CKShareMetadata_GetPropParticipantRole(HandleRef ptr);
         
+
         #endregion
 
         internal CKShareMetadata(IntPtr ptr) : base(ptr) {}
         
-        #region Class Methods
-        
-        #endregion
-
-        #region Constructors
-        
-        #endregion
-
-
-        #region Methods
         
         
-        #endregion
+        
 
-        #region Properties
+
+        
+        
         
         public string ContainerIdentifier 
         {
@@ -182,8 +173,9 @@ namespace HovelHouse.CloudKit
             }
         }
         
-        #endregion
+
         
+
         
         #region IDisposable Support
         #if UNITY_IPHONE || UNITY_TVOS
@@ -195,10 +187,7 @@ namespace HovelHouse.CloudKit
             
         private bool disposedValue = false; // To detect redundant calls
         
-        // No base.Dispose() needed
-        // All we ever do is decrement the reference count in managed code
-        
-        private void Dispose(bool disposing)
+        protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
             {