+++ Runtime/Plugin/NSPredicate.cs
@@ -1,7 +1,7 @@
 //
 //  NSPredicate.cs
 //
-//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/02/2020
+//  Created by Jonathan Culp <jonathanculp@gmail.com> on 03/13/2020
 //  Copyright © 2020 HovelHouseApps. All rights reserved.
 //  Unauthorized copying of this file, via any medium is strictly prohibited
 //  Proprietary and confidential
@@ -42,10 +42,8 @@ namespace HovelHouse.CloudKit
             out IntPtr exceptionPtr);
         
 
-        // Constructors
         
 
-        // Instance Methods
         
 
         
@@ -59,11 +57,11 @@ namespace HovelHouse.CloudKit
         #endif
         private static extern IntPtr NSPredicate_GetPropPredicateFormat(HandleRef ptr);
         
+
         #endregion
 
         internal NSPredicate(IntPtr ptr) : base(ptr) {}
         
-        #region Class Methods
         
         
         public static NSPredicate PredicateWithValue(
@@ -105,19 +103,12 @@ namespace HovelHouse.CloudKit
         
 
         
-        #endregion
-
-        #region Constructors
-        
-        #endregion
-
-
-        #region Methods
         
         
-        #endregion
 
-        #region Properties
+
+        
+        
         
         public string PredicateFormat 
         {
@@ -128,8 +119,9 @@ namespace HovelHouse.CloudKit
             }
         }
         
-        #endregion
+
         
+
         
         #region IDisposable Support
         #if UNITY_IPHONE || UNITY_TVOS
@@ -141,10 +133,7 @@ namespace HovelHouse.CloudKit
             
         private bool disposedValue = false; // To detect redundant calls
         
-        // No base.Dispose() needed
-        // All we ever do is decrement the reference count in managed code
-        
-        private void Dispose(bool disposing)
+        protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
             {